Moved login and logout function to a module of their own
